edge/requirements.txt
flask==3.0.0
scikit-learn==1.3.2
joblib==1.3.2
numpy==1.26.2
psutil==5.9.6
tensorflow-cpu==2.15.0
